+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 1.15 (development)
 
 * Fix: Make sure dkjson decodes "null" to LuaXP's *null* value rather than Lua `nil`.
+* Fix: Some versions of Luasec require the SSL protocol to be set, they don't default, so use "tlsv1" as a default for all.
+* Enhancement: Load and create recipes. A recipe is a saved/packaged configuration that can be deployed easily.
 * Enhancement: Add `SetEnabled` action (parameter `newEnabledValue`) to set enable/disable state of a SiteSensor. When disabled, queries are not run. This provides an alternate mechanism to the arm/disarm control to control queries without Luup notifications that accompany security sensor arming state. Requested by whyfseeguy.
 
 ## Version 1.14 (released)